Separate lua bindings for iso_fs into new file.
Use th_lua_* convention that other classes bound to lua use.
