+++ CorsixTH/Src/iso_fs.cpp
@@ -21,7 +21,6 @@ SOFTWARE.
 */
 
 #include "iso_fs.h"
-#include "th_lua_internal.h"
 #include "th.h"
 #include <cstring>
 #include <cstdarg>
@@ -574,122 +573,4 @@ void iso_filesystem::set_error(const char* sFormat, ...)
     va_start(a, sFormat);
     std::vsnprintf(error, 1024, sFormat, a);
     va_end(a);
-}
-
-namespace {
-
-int l_isofs_new(lua_State *L)
-{
-    luaT_stdnew<iso_filesystem>(L, luaT_environindex, true);
-    return 1;
-}
-
-int l_isofs_set_path_separator(lua_State *L)
-{
-    iso_filesystem *pSelf = luaT_testuserdata<iso_filesystem>(L);
-    pSelf->set_path_separator(luaL_checkstring(L, 2)[0]);
-    lua_settop(L, 1);
-    return 1;
-}
-
-int l_isofs_set_root(lua_State *L)
-{
-    iso_filesystem *pSelf = luaT_testuserdata<iso_filesystem>(L);
-    FILE *fIso = *luaT_testuserdata<FILE*>(L, 2, false);
-    if(pSelf->initialise(fIso))
-    {
-        lua_pushvalue(L, 2);
-        luaT_setenvfield(L, 1, "file");
-        lua_settop(L, 1);
-        return 1;
-    }
-    else
-    {
-        lua_pushnil(L);
-        lua_pushstring(L, pSelf->get_error());
-        return 2;
-    }
-}
-
-int l_isofs_file_exists(lua_State *L)
-{
-    iso_filesystem *pSelf = luaT_testuserdata<iso_filesystem>(L);
-    const char* sFilename = luaL_checkstring(L, 2);
-    iso_filesystem::file_handle iFile = pSelf->find_file(sFilename);
-    if(!iso_filesystem::isHandleGood(iFile))
-    {
-        lua_pushnil(L);
-        lua_pushfstring(L, "Could not find \'%s\' in .iso image", sFilename);
-        return 2;
-    }
-    lua_pushboolean(L, true);
-    return 1;
-}
-
-int l_isofs_file_size(lua_State *L)
-{
-    iso_filesystem *pSelf = luaT_testuserdata<iso_filesystem>(L);
-    const char* sFilename = luaL_checkstring(L, 2);
-    iso_filesystem::file_handle iFile = pSelf->find_file(sFilename);
-    if(!iso_filesystem::isHandleGood(iFile))
-    {
-        lua_pushnil(L);
-        lua_pushfstring(L, "Could not find \'%s\' in .iso image", sFilename);
-        return 2;
-    }
-    lua_pushinteger(L, pSelf->get_file_size(iFile));
-    return 1;
-}
-
-int l_isofs_read_contents(lua_State *L)
-{
-    iso_filesystem *pSelf = luaT_testuserdata<iso_filesystem>(L);
-    const char* sFilename = luaL_checkstring(L, 2);
-    iso_filesystem::file_handle iFile = pSelf->find_file(sFilename);
-    if(!iso_filesystem::isHandleGood(iFile))
-    {
-        lua_pushnil(L);
-        lua_pushfstring(L, "Could not find \'%s\' in .iso image", sFilename);
-        return 2;
-    }
-    void* pBuffer = lua_newuserdata(L, pSelf->get_file_size(iFile));
-    if(!pSelf->get_file_data(iFile, reinterpret_cast<uint8_t*>(pBuffer)))
-    {
-        lua_pushnil(L);
-        lua_pushstring(L, pSelf->get_error());
-        return 2;
-    }
-    lua_pushlstring(L, reinterpret_cast<char*>(pBuffer), pSelf->get_file_size(iFile));
-    return 1;
-}
-
-void l_isofs_list_files_callback(void *p, const char* name, const char* path)
-{
-    lua_State *L = reinterpret_cast<lua_State*>(p);
-    lua_pushstring(L, name);
-    lua_pushstring(L, path);
-    lua_settable(L, 3);
-}
-
-int l_isofs_list_files(lua_State *L)
-{
-    iso_filesystem *pSelf = luaT_testuserdata<iso_filesystem>(L);
-    const char* sPath = luaL_checkstring(L, 2);
-    lua_settop(L, 2);
-    lua_newtable(L);
-    pSelf->visit_directory_files(sPath, l_isofs_list_files_callback, L);
-    return 1;
-}
-
-} // namespace
-
-void lua_register_iso_fs(const lua_register_state* pState)
-{
-    lua_class_binding<iso_filesystem> lcb(pState, "iso_fs", l_isofs_new, lua_metatable::iso_fs);
-    lcb.add_function(l_isofs_set_path_separator, "setPathSeparator");
-    lcb.add_function(l_isofs_set_root, "setRoot");
-    lcb.add_function(l_isofs_file_exists, "fileExists");
-    lcb.add_function(l_isofs_file_size, "fileSize");
-    lcb.add_function(l_isofs_read_contents, "readContents");
-    lcb.add_function(l_isofs_list_files, "listFiles");
 }
